Update on "[BE][Dynamo]Add silent incorrectness fix"
  ---                                                                                                                                                                  
  Fix silent incorrectness issues in dynamo modules                                                                                                                    
                                                                                                                                                                       
  Summary                                                                                                                                                              
                                                                                                                                                                       
  This PR addresses several potential silent incorrectness issues in the dynamo modules where code could run without errors but produce wrong results:                 
                                                                                                                                                                       
  1. Fixed weakref comparison in cache validation (cache_size.py): The _has_same_id_matched_objs function was comparing weakref container objects with is instead of   
  dereferencing them first. This could cause incorrect cache reuse when different weakref objects reference the same underlying object.                                
  2. Fixed falsy cell contents handling (side_effects.py): The load_cell method used a truthiness check (if cellvar.pre_existing_contents:) which incorrectly treated  
  valid falsy values like 0, False, or "" as uninitialized cells. Changed to explicit is not None check.                                                               
  3. Added dead watcher cleanup (mutation_guard.py): The MutationTracker.track() method could accumulate dead weak references unboundedly. Added periodic cleanup when 
  the watcher list exceeds 100 entries.                                                                                                                                
  4. Added mutation ignore logging (side_effects.py): Added debug logging to ignore_mutations_on and stop_ignoring_mutations_on to help diagnose issues where mutations
   are silently lost.                                                                                                                                                  
  5. Fixed falsy VariableTracker handling (dicts.py): The is_new_item method used a truthiness check that could skip realization comparison for falsy VariableTrackers.
  6. Added defensive None check (side_effects.py): The is_modified method had an assert that would crash if mutation_type was unexpectedly None. Changed to return     
  False defensively.                                                                                                                                                   
                                                                                                                                                                       
  Test plan                                                                                                                                                            
                                                                                                                                                                       
  - Added new test file test/dynamo/test_silent_correctness_fixes.py with 12 unit tests covering all fixes                                                             
  - All new tests pass                                                                                                                                                 
  - Smoke tests for basic dynamo compilation and mutation tracking pass                                                                                                
                                                                                                                                                                       
  python test/dynamo/test_silent_correctness_fixes.py -v                                                                                                               
                                                                                                                                                                       
  ---                                                                                                                                                                  
  This PR was authored with assistance from Claude.  




cc voznesenskym penguinwu EikanWang jgong5 Guobing-Chen XiaobingSuper zhuhaozhe blzheng wenzhe-nrv jiayisunx kadeng chauhang amjames Lucaskabela jataylo

[ghstack-poisoned]
